第7回 Lesson1 サーバサイド

+++ app/Controller/TodoListsController.php
@@ -3,18 +3,19 @@
 App::uses('AppController', 'Controller');
 
 class TodoListsController extends AppController {
+	private $fields = array (
+		'TodoList.id',
+		'TodoList.todo',
+		'TodoList.status',
+		'Owner.id',
+		'Owner.name',
+		'Assignee.id',
+		'Assignee.name'
+	);
 
 	public function index() {
 		$query = array (
-			'fields' => array (
-				'TodoList.id',
-				'TodoList.todo',
-				'TodoList.status',
-				'Owner.id',
-				'Owner.name',
-				'Assignee.id',
-				'Assignee.name'
-			),
+			'fields' => $this->fields,
 			'order' => "TodoList.id"
 		);
 		$res = $this->TodoList->find('all', $query);
@@ -33,7 +34,7 @@ class TodoListsController extends AppController {
 	}
 
 	public function view($id = null) {
-		$res = $this->TodoList->findById($id);
+		$res = $this->TodoList->findById($id, $this->fields);
 		$this->set(compact('res'));
 		$this->set('_serialize', 'res');
 	}
@@ -42,13 +43,21 @@ class TodoListsController extends AppController {
 		$data = $this->request->data;
 		$data['owner'] = $this->Auth->user()['id'];
 		$res = $this->TodoList->save($data);
-		$this->set(compact('res'));
-		$this->set('_serialize', 'res');
+		$response = $this->editResponse($res);
+		$this->set(compact('response'));
+		$this->set('_serialize', 'response');
 	}
 	public function delete($id) {
-		$res = $this->TodoList->delete($id, false);
-		$this->set(compact('res'));
-		$this->set('_serialize', 'res');
+		//オーナかどうかチェック
+		if(!$this->TodoList->isOwner($id)){
+			$this->setStatusValidationError();
+			$response = $this->editErrors('オーナのみ削除可能です。');
+		}else{
+			$res = $this->TodoList->delete($id, false);
+			$response = $this->editResponse($res);
+		}
+		$this->set(compact('response'));
+		$this->set('_serialize', 'response');
 	}
 
 	public function edit($id) {
@@ -56,8 +65,40 @@ class TodoListsController extends AppController {
 		$data = $this->request->data;
 		$res = $this->TodoList->save($this->request->data);
 		$res = !empty($res);
-		$this->set(compact('res'));
-		$this->set('_serialize', 'res');
+		$response = $this->editResponse($res);
+		$this->set(compact('response'));
+		$this->set('_serialize', 'response');
+	}
+
+	//レスポンスを編集
+	private function editResponse($res){
+		if($res){
+			$response = $res;
+		}else{
+			$this->setStatusValidationError();
+			$respnse = array();
+			if(count($this->TodoList->validationErrors) > 0){
+				$response = $this->editErrors($this->TodoList->validationErrors);
+			}else{
+				$response = $this->editErrors('エラーが発生しました。');
+			}
+		}
+		return $response;
+	}
+
+	//バリデーションエラー時はレスポンスを400に設定
+	private function setStatusValidationError(){
+		$this->response->statusCode(400);
+	}
+
+	//エラーメッセージを編集
+	private function editErrors($errors){
+		if(is_array($errors)){
+			$res['errors'] = $errors;
+		}else{
+			$res['errors']  = array('error' => array($errors));
+		}
+		return $res;
 	}
 
 }
